Repos/EsportW/requirements.txt
plyer==2.1.0
PyYAML==6.0
requests==2.28.2
retrying==1.3.4
rich==13.3.3
selenium==4.8.3
undetected_chromedriver==3.4.6
urllib3==1.26.15
webdriver_manager==3.8.5
